backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
google-genai>=1.0.0
pydantic>=2.7.1
httpx>=0.27.0
python-multipart>=0.0.9
networkx>=3.3
pytest>=8.2.0
